+++ CHANGELOG.md
@@ -5,10 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is loosely based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-Parsley repects [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-Minor patches are placed in their own directories. Patches are bucketed into the
-directory of their minor version.
-
 ## [0.1.0] - 2021-04-09
 
 ### Added
